refactor: change travel add edit file to ride
